added all files for a working site for users to list items and delete. Bug fixes to follow

+++ controllers/categoryCtrl.js
@@ -1,4 +1,5 @@
 const Category = require ('../models/categoryModel')
+const Products = require ('../models/ProductModel')
 
 const categoryCtrl = {
     getCategories: async (req, res) => {
@@ -27,6 +28,11 @@ const categoryCtrl = {
     },
     deleteCategory: async(req, res) => {
         try {
+            const products = await Products.findOne({category: req.params.id})
+            if(products) return res.status(400).json({
+                msg: "This category contains active products, please delete these first"
+            })
+
             await Category.findByIdAndDelete(req.params.id)
             res.json({msg: "Deleted a category"})
         } catch (err) {
